@@ -19,7 +19,7 @@ package bean
 
 import (
 	"encoding/json"
-	"github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
+	"github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 )
 
 type AppContainer struct {
