@@ -21,18 +21,20 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/argoproj/gitops-engine/pkg/health"
 	client "github.com/devtron-labs/devtron/api/helm-app"
 	"github.com/devtron-labs/devtron/api/restHandler/common"
 	"github.com/devtron-labs/devtron/pkg/cluster"
 	"github.com/devtron-labs/devtron/pkg/user/casbin"
 	util2 "github.com/devtron-labs/devtron/util"
+	"github.com/devtron-labs/devtron/util/argo"
 	"net/http"
 	"strconv"
 	"strings"
 	"time"
 
-	application2 "github.com/argoproj/argo-cd/pkg/apiclient/application"
-	"github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
+	application2 "github.com/argoproj/argo-cd/v2/pkg/apiclient/application"
+	"github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 	"github.com/devtron-labs/devtron/api/bean"
 	"github.com/devtron-labs/devtron/client/argocdServer/application"
 	"github.com/devtron-labs/devtron/internal/constants"
@@ -71,6 +73,7 @@ type AppListingRestHandlerImpl struct {
 	helmAppClient          client.HelmAppClient
 	clusterService         cluster.ClusterService
 	helmAppService         client.HelmAppService
+	argoUserService        argo.ArgoUserService
 }
 
 type AppStatus struct {
@@ -88,7 +91,8 @@ func NewAppListingRestHandlerImpl(application application.ServiceClient,
 	pipeline pipeline.PipelineBuilder,
 	logger *zap.SugaredLogger, enforcerUtil rbac.EnforcerUtil,
 	deploymentGroupService deploymentGroup.DeploymentGroupService, userService user.UserService,
-	helmAppClient client.HelmAppClient, clusterService cluster.ClusterService, helmAppService client.HelmAppService) *AppListingRestHandlerImpl {
+	helmAppClient client.HelmAppClient, clusterService cluster.ClusterService, helmAppService client.HelmAppService,
+	argoUserService argo.ArgoUserService) *AppListingRestHandlerImpl {
 	appListingHandler := &AppListingRestHandlerImpl{
 		application:            application,
 		appListingService:      appListingService,
@@ -102,6 +106,7 @@ func NewAppListingRestHandlerImpl(application application.ServiceClient,
 		helmAppClient:          helmAppClient,
 		clusterService:         clusterService,
 		helmAppService:         helmAppService,
+		argoUserService:        argoUserService,
 	}
 	return appListingHandler
 }
@@ -325,7 +330,13 @@ func (handler AppListingRestHandlerImpl) FetchAppTriggerView(w http.ResponseWrit
 			}
 		}(ctx.Done(), cn.CloseNotify())
 	}
-	ctx = context.WithValue(ctx, "token", token)
+	acdToken, err := handler.argoUserService.GetLatestDevtronArgoCdUserToken()
+	if err != nil {
+		handler.logger.Errorw("error in getting acd token", "err", err)
+		common.WriteJsonResp(w, err, nil, http.StatusInternalServerError)
+		return
+	}
+	ctx = context.WithValue(ctx, "token", acdToken)
 	defer cancel()
 
 	response := make(chan AppStatus)
@@ -354,7 +365,7 @@ func (handler AppListingRestHandlerImpl) FetchAppTriggerView(w http.ResponseWrit
 					healthStatus := resp.Application.Status.Health.Status
 					status := AppStatus{
 						name:       pipelineName,
-						status:     healthStatus,
+						status:     string(healthStatus),
 						message:    resp.Application.Status.Health.Message,
 						err:        nil,
 						conditions: resp.Application.Status.Conditions,
@@ -403,7 +414,7 @@ func (handler AppListingRestHandlerImpl) FetchAppTriggerView(w http.ResponseWrit
 		if triggerView[i].Status == "" {
 			triggerView[i].Status = "Unknown"
 		}
-		if triggerView[i].Status == v1alpha1.HealthStatusDegraded {
+		if triggerView[i].Status == string(health.HealthStatusDegraded) {
 			triggerView[i].Status = "Not Deployed"
 		}
 	}
@@ -544,8 +555,14 @@ func (handler AppListingRestHandlerImpl) fetchResourceTree(w http.ResponseWriter
 				}
 			}(ctx.Done(), cn.CloseNotify())
 		}
-		ctx = context.WithValue(ctx, "token", token)
 		defer cancel()
+		acdToken, err := handler.argoUserService.GetLatestDevtronArgoCdUserToken()
+		if err != nil {
+			handler.logger.Errorw("error in getting acd token", "err", err)
+			common.WriteJsonResp(w, err, "", http.StatusInternalServerError)
+			return appDetail
+		}
+		ctx = context.WithValue(ctx, "token", acdToken)
 		start := time.Now()
 		resp, err := handler.application.ResourceTree(ctx, query)
 		elapsed := time.Since(start)
@@ -559,7 +576,7 @@ func (handler AppListingRestHandlerImpl) fetchResourceTree(w http.ResponseWriter
 			common.WriteJsonResp(w, err, "", http.StatusInternalServerError)
 			return appDetail
 		}
-		if resp.Status == v1alpha1.HealthStatusHealthy {
+		if resp.Status == string(health.HealthStatusHealthy) {
 			status, err := handler.appListingService.ISLastReleaseStopType(appId, envId)
 			if err != nil {
 				handler.logger.Errorw("service err, FetchAppDetails", "err", err, "app", appId, "env", envId)
@@ -569,7 +586,7 @@ func (handler AppListingRestHandlerImpl) fetchResourceTree(w http.ResponseWriter
 		}
 		handler.logger.Debugw("FetchAppDetails, time elapsed in fetching application for environment ", "elapsed", elapsed, "appId", appId, "envId", envId)
 
-		if resp.Status == v1alpha1.HealthStatusDegraded {
+		if resp.Status == string(health.HealthStatusDegraded) {
 			count, err := handler.appListingService.GetReleaseCount(appId, envId)
 			if err != nil {
 				handler.logger.Errorw("service err, FetchAppDetails, release count", "err", err, "app", appId, "env", envId)
