@@ -23,6 +23,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/devtron-labs/devtron/pkg/chart"
+	"github.com/devtron-labs/devtron/util/argo"
 	"io"
 	"net/http"
 	"strconv"
@@ -111,6 +112,7 @@ type PipelineConfigRestHandlerImpl struct {
 	policyService           security2.PolicyService
 	scanResultRepository    security.ImageScanResultRepository
 	gitProviderRepo         repository.GitProviderRepository
+	argoUserService         argo.ArgoUserService
 }
 
 func NewPipelineRestHandlerImpl(pipelineBuilder pipeline.PipelineBuilder, Logger *zap.SugaredLogger,
@@ -131,7 +133,8 @@ func NewPipelineRestHandlerImpl(pipelineBuilder pipeline.PipelineBuilder, Logger
 	appCloneService appClone.AppCloneService,
 	appWorkflowService appWorkflow.AppWorkflowService,
 	materialRepository pipelineConfig.MaterialRepository, policyService security2.PolicyService,
-	scanResultRepository security.ImageScanResultRepository, gitProviderRepo repository.GitProviderRepository) *PipelineConfigRestHandlerImpl {
+	scanResultRepository security.ImageScanResultRepository, gitProviderRepo repository.GitProviderRepository,
+	argoUserService argo.ArgoUserService) *PipelineConfigRestHandlerImpl {
 	return &PipelineConfigRestHandlerImpl{
 		pipelineBuilder:         pipelineBuilder,
 		Logger:                  Logger,
@@ -158,6 +161,7 @@ func NewPipelineRestHandlerImpl(pipelineBuilder pipeline.PipelineBuilder, Logger
 		policyService:           policyService,
 		scanResultRepository:    scanResultRepository,
 		gitProviderRepo:         gitProviderRepo,
+		argoUserService:         argoUserService,
 	}
 }
 
@@ -260,7 +264,13 @@ func (handler PipelineConfigRestHandlerImpl) CreateApp(w http.ResponseWriter, r 
 				}
 			}(ctx.Done(), cn.CloseNotify())
 		}
-		ctx = context.WithValue(r.Context(), "token", token)
+		acdToken, err := handler.argoUserService.GetLatestDevtronArgoCdUserToken()
+		if err != nil {
+			handler.Logger.Errorw("error in getting acd token", "err", err)
+			common.WriteJsonResp(w, err, nil, http.StatusInternalServerError)
+			return
+		}
+		ctx = context.WithValue(r.Context(), "token", acdToken)
 		createResp, err = handler.appCloneService.CloneApp(&createRequest, ctx)
 	}
 	if err != nil {
