@@ -20,8 +20,8 @@ package application
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
-	"github.com/argoproj/argo-cd/util/settings"
+	"github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
+	"github.com/argoproj/argo-cd/v2/util/settings"
 	"testing"
 
 	"go.uber.org/zap"
