@@ -1,6 +1,7 @@
 package application
 
 import (
+	"context"
 	"encoding/json"
 	"github.com/devtron-labs/devtron/pkg/cluster/repository"
 	"go.uber.org/zap"
@@ -74,9 +75,9 @@ func (impl K8sClientServiceImpl) GetResource(restConfig *rest.Config, request *K
 	resourceIdentifier := request.ResourceIdentifier
 	var resp *unstructured.Unstructured
 	if len(resourceIdentifier.Namespace) > 0 && namespaced {
-		resp, err = resourceIf.Namespace(resourceIdentifier.Namespace).Get(resourceIdentifier.Name, metav1.GetOptions{})
+		resp, err = resourceIf.Namespace(resourceIdentifier.Namespace).Get(context.Background(), resourceIdentifier.Name, metav1.GetOptions{})
 	} else {
-		resp, err = resourceIf.Get(resourceIdentifier.Name, metav1.GetOptions{})
+		resp, err = resourceIf.Get(context.Background(), resourceIdentifier.Name, metav1.GetOptions{})
 	}
 	if err != nil {
 		impl.logger.Errorw("error in getting resource", "err", err, "resource", resourceIdentifier.Name)
@@ -100,9 +101,9 @@ func (impl K8sClientServiceImpl) CreateResource(restConfig *rest.Config, request
 	resourceIdentifier := request.ResourceIdentifier
 	var resp *unstructured.Unstructured
 	if len(resourceIdentifier.Namespace) > 0 && namespaced {
-		resp, err = resourceIf.Namespace(resourceIdentifier.Namespace).Create(&unstructured.Unstructured{Object: createObj}, metav1.CreateOptions{})
+		resp, err = resourceIf.Namespace(resourceIdentifier.Namespace).Create(context.Background(), &unstructured.Unstructured{Object: createObj}, metav1.CreateOptions{})
 	} else {
-		resp, err = resourceIf.Create(&unstructured.Unstructured{Object: createObj}, metav1.CreateOptions{})
+		resp, err = resourceIf.Create(context.Background(), &unstructured.Unstructured{Object: createObj}, metav1.CreateOptions{})
 	}
 	if err != nil {
 		impl.logger.Errorw("error in creating resource", "err", err)
@@ -126,9 +127,9 @@ func (impl K8sClientServiceImpl) UpdateResource(restConfig *rest.Config, request
 	resourceIdentifier := request.ResourceIdentifier
 	var resp *unstructured.Unstructured
 	if len(resourceIdentifier.Namespace) > 0 && namespaced {
-		resp, err = resourceIf.Namespace(resourceIdentifier.Namespace).Update(&unstructured.Unstructured{Object: updateObj}, metav1.UpdateOptions{})
+		resp, err = resourceIf.Namespace(resourceIdentifier.Namespace).Update(context.Background(), &unstructured.Unstructured{Object: updateObj}, metav1.UpdateOptions{})
 	} else {
-		resp, err = resourceIf.Update(&unstructured.Unstructured{Object: updateObj}, metav1.UpdateOptions{})
+		resp, err = resourceIf.Update(context.Background(), &unstructured.Unstructured{Object: updateObj}, metav1.UpdateOptions{})
 	}
 	if err != nil {
 		impl.logger.Errorw("error in updating resource", "err", err, "resource", resourceIdentifier.Name)
@@ -145,19 +146,19 @@ func (impl K8sClientServiceImpl) DeleteResource(restConfig *rest.Config, request
 	resourceIdentifier := request.ResourceIdentifier
 	var obj *unstructured.Unstructured
 	if len(resourceIdentifier.Namespace) > 0 && namespaced {
-		obj, err = resourceIf.Namespace(resourceIdentifier.Namespace).Get(request.ResourceIdentifier.Name, metav1.GetOptions{})
+		obj, err = resourceIf.Namespace(resourceIdentifier.Namespace).Get(context.Background(), request.ResourceIdentifier.Name, metav1.GetOptions{})
 		if err != nil {
 			impl.logger.Errorw("error in getting resource", "err", err, "resource", resourceIdentifier.Name)
 			return nil, err
 		}
-		err = resourceIf.Namespace(resourceIdentifier.Namespace).Delete(request.ResourceIdentifier.Name, &metav1.DeleteOptions{})
+		err = resourceIf.Namespace(resourceIdentifier.Namespace).Delete(context.Background(), request.ResourceIdentifier.Name, metav1.DeleteOptions{})
 	} else {
-		obj, err = resourceIf.Get(request.ResourceIdentifier.Name, metav1.GetOptions{})
+		obj, err = resourceIf.Get(context.Background(), request.ResourceIdentifier.Name, metav1.GetOptions{})
 		if err != nil {
 			impl.logger.Errorw("error in getting resource", "err", err, "resource", resourceIdentifier.Name)
 			return nil, err
 		}
-		err = resourceIf.Delete(request.ResourceIdentifier.Name, &metav1.DeleteOptions{})
+		err = resourceIf.Delete(context.Background(), request.ResourceIdentifier.Name, metav1.DeleteOptions{})
 	}
 	if err != nil {
 		impl.logger.Errorw("error in deleting resource", "err", err, "resource", resourceIdentifier.Name)
@@ -193,7 +194,7 @@ func (impl K8sClientServiceImpl) ListEvents(restConfig *rest.Config, request *K8
 		},
 		FieldSelector: fieldSelector.String(),
 	}
-	list, err := eventsIf.List(listOptions)
+	list, err := eventsIf.List(context.Background(), listOptions)
 	if err != nil {
 		impl.logger.Errorw("error in getting events list", "err", err)
 		return nil, err
@@ -221,7 +222,7 @@ func (impl K8sClientServiceImpl) GetPodLogs(restConfig *rest.Config, request *K8
 	}
 	podIf := podClient.Pods(resourceIdentifier.Namespace)
 	logsRequest := podIf.GetLogs(resourceIdentifier.Name, podLogOptions)
-	stream, err := logsRequest.Stream()
+	stream, err := logsRequest.Stream(context.Background())
 	if err != nil {
 		impl.logger.Errorw("error in streaming pod logs", "err", err)
 		return nil, err
