@@ -19,7 +19,7 @@ package ArgoUtil
 
 import (
 	"context"
-	"github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
+	"github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 )
 
 type ApplicationService interface {
