@@ -18,7 +18,7 @@
 package ArgoUtil
 
 import (
-	"github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
+	"github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 )
 
 type ResourceService interface {
