@@ -18,6 +18,7 @@
 package util
 
 import (
+	"context"
 	"encoding/json"
 	error2 "errors"
 	"flag"
@@ -165,7 +166,7 @@ func (impl K8sUtil) CreateNsIfNotExists(namespace string, clusterConfig *Cluster
 }
 
 func (impl K8sUtil) checkIfNsExists(namespace string, client *v12.CoreV1Client) (exists bool, err error) {
-	ns, err := client.Namespaces().Get(namespace, metav1.GetOptions{})
+	ns, err := client.Namespaces().Get(context.Background(), namespace, metav1.GetOptions{})
 	//ns, err := impl.k8sClient.CoreV1().Namespaces().Get(namespace, metav1.GetOptions{})
 	impl.logger.Debugw("ns fetch", "name", namespace, "res", ns)
 	if errors.IsNotFound(err) {
@@ -180,7 +181,7 @@ func (impl K8sUtil) checkIfNsExists(namespace string, client *v12.CoreV1Client) 
 
 func (impl K8sUtil) createNs(namespace string, client *v12.CoreV1Client) (ns *v1.Namespace, err error) {
 	nsSpec := &v1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
-	ns, err = client.Namespaces().Create(nsSpec)
+	ns, err = client.Namespaces().Create(context.Background(), nsSpec, metav1.CreateOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -189,12 +190,12 @@ func (impl K8sUtil) createNs(namespace string, client *v12.CoreV1Client) (ns *v1
 }
 
 func (impl K8sUtil) deleteNs(namespace string, client *v12.CoreV1Client) error {
-	err := client.Namespaces().Delete(namespace, &metav1.DeleteOptions{})
+	err := client.Namespaces().Delete(context.Background(), namespace, metav1.DeleteOptions{})
 	return err
 }
 
 func (impl K8sUtil) GetConfigMap(namespace string, name string, client *v12.CoreV1Client) (*v1.ConfigMap, error) {
-	cm, err := client.ConfigMaps(namespace).Get(name, metav1.GetOptions{})
+	cm, err := client.ConfigMaps(namespace).Get(context.Background(), name, metav1.GetOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -203,7 +204,7 @@ func (impl K8sUtil) GetConfigMap(namespace string, name string, client *v12.Core
 }
 
 func (impl K8sUtil) CreateConfigMap(namespace string, cm *v1.ConfigMap, client *v12.CoreV1Client) (*v1.ConfigMap, error) {
-	cm, err := client.ConfigMaps(namespace).Create(cm)
+	cm, err := client.ConfigMaps(namespace).Create(context.Background(), cm, metav1.CreateOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -212,7 +213,7 @@ func (impl K8sUtil) CreateConfigMap(namespace string, cm *v1.ConfigMap, client *
 }
 
 func (impl K8sUtil) UpdateConfigMap(namespace string, cm *v1.ConfigMap, client *v12.CoreV1Client) (*v1.ConfigMap, error) {
-	cm, err := client.ConfigMaps(namespace).Update(cm)
+	cm, err := client.ConfigMaps(namespace).Update(context.Background(), cm, metav1.UpdateOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -229,7 +230,7 @@ func (impl K8sUtil) PatchConfigMap(namespace string, clusterConfig *ClusterConfi
 	if err != nil {
 		panic(err)
 	}
-	cm, err := client.ConfigMaps(namespace).Patch(name, types.PatchType(types.MergePatchType), b)
+	cm, err := client.ConfigMaps(namespace).Patch(context.Background(), name, types.PatchType(types.MergePatchType), b, metav1.PatchOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -255,7 +256,7 @@ func (impl K8sUtil) PatchConfigMapJsonType(namespace string, clusterConfig *Clus
 		panic(err)
 	}
 
-	cm, err := client.ConfigMaps(namespace).Patch(name, types.PatchType(types.JSONPatchType), b)
+	cm, err := client.ConfigMaps(namespace).Patch(context.Background(), name, types.PatchType(types.JSONPatchType), b, metav1.PatchOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -271,7 +272,7 @@ type JsonPatchType struct {
 }
 
 func (impl K8sUtil) GetSecret(namespace string, name string, client *v12.CoreV1Client) (*v1.Secret, error) {
-	secret, err := client.Secrets(namespace).Get(name, metav1.GetOptions{})
+	secret, err := client.Secrets(namespace).Get(context.Background(), name, metav1.GetOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -286,7 +287,7 @@ func (impl K8sUtil) CreateSecret(namespace string, data map[string][]byte, secre
 		},
 		Data: data,
 	}
-	secret, err := client.Secrets(namespace).Create(secret)
+	secret, err := client.Secrets(namespace).Create(context.Background(), secret, metav1.CreateOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -295,7 +296,7 @@ func (impl K8sUtil) CreateSecret(namespace string, data map[string][]byte, secre
 }
 
 func (impl K8sUtil) UpdateSecret(namespace string, secret *v1.Secret, client *v12.CoreV1Client) (*v1.Secret, error) {
-	secret, err := client.Secrets(namespace).Update(secret)
+	secret, err := client.Secrets(namespace).Update(context.Background(), secret, metav1.UpdateOptions{})
 	if err != nil {
 		return nil, err
 	} else {
@@ -311,14 +312,14 @@ func (impl K8sUtil) DeleteJob(namespace string, name string, clusterConfig *Clus
 	}
 	jobs := clientSet.BatchV1().Jobs(namespace)
 
-	job, err := jobs.Get(name, metav1.GetOptions{})
+	job, err := jobs.Get(context.Background(), name, metav1.GetOptions{})
 	if err != nil && errors.IsNotFound(err) {
 		impl.logger.Errorw("get job err, DeleteJob", "err", err)
 		return nil
 	}
 
 	if job != nil {
-		err := jobs.Delete(name, &metav1.DeleteOptions{})
+		err := jobs.Delete(context.Background(), name, metav1.DeleteOptions{})
 		if err != nil && !errors.IsNotFound(err) {
 			impl.logger.Errorw("delete err, DeleteJob", "err", err)
 			return err
@@ -336,17 +337,17 @@ func (impl K8sUtil) CreateJob(namespace string, name string, clusterConfig *Clus
 	time.Sleep(5 * time.Second)
 
 	jobs := clientSet.BatchV1().Jobs(namespace)
-	_, err = jobs.Get(name, metav1.GetOptions{})
+	_, err = jobs.Get(context.Background(), name, metav1.GetOptions{})
 	if err == nil {
 		impl.logger.Errorw("get job err, CreateJob", "err", err)
 		time.Sleep(5 * time.Second)
-		_, err = jobs.Get(name, metav1.GetOptions{})
+		_, err = jobs.Get(context.Background(), name, metav1.GetOptions{})
 		if err == nil {
 			return error2.New("job deletion takes more time than expected, please try after sometime")
 		}
 	}
 
-	_, err = jobs.Create(job)
+	_, err = jobs.Create(context.Background(), job, metav1.CreateOptions{})
 	if err != nil {
 		impl.logger.Errorw("create err, CreateJob", "err", err)
 		return err
@@ -368,7 +369,7 @@ func (impl K8sUtil) DeletePodByLabel(namespace string, labels string, clusterCon
 	time.Sleep(2 * time.Second)
 
 	pods := clientSet.CoreV1().Pods(namespace)
-	podList, err := pods.List(metav1.ListOptions{LabelSelector: labels})
+	podList, err := pods.List(context.Background(), metav1.ListOptions{LabelSelector: labels})
 	if err != nil && errors.IsNotFound(err) {
 		impl.logger.Errorw("get pod err, DeletePod", "err", err)
 		return nil
@@ -377,7 +378,7 @@ func (impl K8sUtil) DeletePodByLabel(namespace string, labels string, clusterCon
 	for _, pod := range (*podList).Items {
 		if pod.Status.Phase != Running {
 			podName := pod.ObjectMeta.Name
-			err := pods.Delete(podName, &metav1.DeleteOptions{})
+			err := pods.Delete(context.Background(), podName, metav1.DeleteOptions{})
 			if err != nil && !errors.IsNotFound(err) {
 				impl.logger.Errorw("delete err, DeletePod", "err", err)
 				return err
@@ -421,7 +422,7 @@ func (impl K8sUtil) DeleteAndCreateJob(content []byte, namespace string, cluster
 }
 
 func (impl K8sUtil) ListNamespaces(client *v12.CoreV1Client) (*v1.NamespaceList, error) {
-	nsList, err := client.Namespaces().List(metav1.ListOptions{})
+	nsList, err := client.Namespaces().List(context.Background(), metav1.ListOptions{})
 	if errors.IsNotFound(err) {
 		return nsList, nil
 	} else if err != nil {
@@ -448,7 +449,7 @@ func (impl K8sUtil) GetResourceInfoByLabelSelector(namespace string, labelSelect
 		impl.logger.Errorw("cluster config error", "err", err)
 		return nil, err
 	}
-	pods, err := client.Pods(namespace).List(metav1.ListOptions{
+	pods, err := client.Pods(namespace).List(context.Background(), metav1.ListOptions{
 		LabelSelector: labelSelector,
 	})
 	if err != nil {
